WIP MacOS target
Added missing platform files, and fixes some minor issues to continue setting up MacOS target.

+++ framework/SceneGraph/Sources/ModuleSceneGraph.cpp
@@ -181,13 +181,13 @@ void ModuleSceneGraph::DoDefferedAction()
 			break;
 		case DefferedAction::DESTROY_TEXTURE:
 		{
-			unsigned int textureID = (unsigned int)(itr->first);
+			unsigned int textureID = (unsigned int)(uintptr_t)(itr->first);
 			myRenderer->DeleteTexture(1, &textureID);
 			break;
 		}
 		case DefferedAction::DESTROY_BUFFER:
 		{	
-			unsigned int buffer = (unsigned int)(itr->first);
+			unsigned int buffer = (unsigned int)(uintptr_t)(itr->first);
 			myRenderer->DeleteBuffer(1, &buffer);
 			break;
 		}
